# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Wallet API
  description: >
    Digital Wallet backend (FinTech intern project). Modular-layered monolith
    exposing user auth, wallet balances, top-up, P2P transfer, merchant payment,
    transaction history, and admin operations. See SRS.md / HLD.md / LLD.md.
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  page:
                    type: integer
                  size:
                    type: integer
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/admin/transactions/{id}/reverse:
    post:
      tags: [Admin]
      summary: Reverse a transaction (admin only)
      description: >
        Creates a compensating REVERSAL transaction and ledger entries,
        restoring balances (FR-8).
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Reversal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TxResult"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Transaction already reversed or not in a reversible state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/users/{id}/disable:
    patch:
      tags: [Admin]
      summary: Toggle user enabled/disabled status (admin only)
      description: >
        Toggles the `enabled` field on a user account. When disabled, the user
        cannot log in. Admin users cannot be disabled.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/admin/users/{id}:
    delete:
      tags: [Admin]
      summary: Delete a user (admin only)
      description: Permanently deletes a user account. Admin users cannot be deleted.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TransactionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not permitted (e.g. non-admin calling an admin route)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    Role:
      type: string
      enum: [USER, MERCHANT, ADMIN]
    TxType:
      type: string
      enum: [TOP_UP, TRANSFER, PAYMENT, REVERSAL]
    TxStatus:
      type: string
      enum: [PENDING, SUCCESS, FAILED, REVERSED]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/TxType"
        amount:
          type: number
          format: double
        status:
          $ref: "#/components/schemas/TxStatus"
        sourceWalletId:
          type: string
          format: uuid
          nullable: true
        targetWalletId:
          type: string
          format: uuid
          nullable: true
        reversedTransactionId:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
    TxResult:
      type: object
      properties:
        transaction:
          $ref: "#/components/schemas/Transaction"
        walletBalance:
          type: number
          format: double
          description: Balance of the acting user's wallet after this operation
...
    AdminUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          example: "INSUFFICIENT_FUNDS"
        message:
          type: string
        correlationId:
          type: string